+++ src/main/kotlin/ris58h/webcalm/javascript/psi/JavaScriptMethodDefinition.kt
@@ -1,0 +1,6 @@
+package ris58h.webcalm.javascript.psi
+
+import com.intellij.extapi.psi.ASTWrapperPsiElement
+import com.intellij.lang.ASTNode
+
+class JavaScriptMethodDefinition(node: ASTNode) : ASTWrapperPsiElement(node)